bugfix in plot_ramanspectrum.py

+++ cube2jvxl.sh
@@ -16,12 +16,42 @@
 # Conver CUBE isosurface data into JVXL format and pack all data into ZIP
 # using JMOL software  (http://jmol.sourceforge.net)
 #
+# Note: In order to save isosurface data one have to put specific jmolscript
+# command inside CUBE input file (second line) For example: 
+# jmolscript: isosurface s1 sign cyan yellow cutoff 1.000000 "" color translucent 0.45
+# or provide such command  with -c parameter. For example:
+# cube2jvxl.sh -i input.cube -c 'isosurface s1 sign cyan yellow cutoff 1.0 "" color translucent 0.45'
+# Notice on single and double quote. If you whant to use only double quotes don't forget to slash it:
+# cube2jvxl.sh -i input.cube -c "isosurface s1 sign cyan yellow cutoff 1.0 \"\" color translucent 0.45"
+# 
+# First string after comments is the path to Jmol executable. Set it according to your system installation.
+#
 # Example how to convert all .cube files in the directory:
 # $ for i in *.cube; do ./cube2jvxl.sh -i $i; done
 #
 # Author: Eugene Roginskii
 
+#Do not modify here, see below
+if [[ -n $JMOLEXE ]]; then
+    Jmolexe=$JMOLEXE
+    echo "Jmol path is set according to JMOLEXE variable"
+else
+# Modify line below to provide full path to Jmol executable
+    Jmolexe=/opt/jmol/jmol.sh
+fi
 
+
+
+if [[ ! -x ${Jmolexe} ]]; then
+    echo ${Jmolexe} not found
+    if [[ -n `which jmol.sh 2>&1 | xargs file  | grep "No such file or directory"` ]]; then
+        echo "Jmol not found. Provide full path to Jmol executable"
+        echo " by setting JMOLEXE global variable in your .bashrc"
+        exit 1
+    else
+        Jmolexe=`which jmol.sh` || exit 1
+    fi
+fi
 
 atom_data=(X H He Li Be B C N O F Ne Na Mg Al Si P S Cl Ar K Ca Sc Ti V Cr Mn Fe Co Ni Cu Zn Ga Ge \
 As Se Br Kr Rb Sr Y Zr Nb Mo Tc Ru Rh Pd Ag Cd In Sn Sb Te I Xe Cs Ba La Ce Pr Nd Pm Sm Eu Gd Tb Dy \
@@ -34,22 +64,39 @@ cat << EOF >&2
 usage: $0 options
 
 Converter CUBE isosurface file format to jvxl one
+Note: In order to save isosurface data one have to put specific jmolscript
+command inside CUBE input file (second line) For example: 
+jmolscript: isosurface s1 sign cyan yellow cutoff 1.000000 "" color translucent 0.45
+or provide such command  with -c parameter. For example:
+cube2jvxl.sh -i input.cube -c 'isosurface s1 sign cyan yellow cutoff 1.0 "" color translucent 0.45'
+Notice on single and double quote. If you whant to use only double quotes don't forget to slash it:
+cube2jvxl.sh -i input.cube -c "isosurface s1 sign cyan yellow cutoff 1.0 \"\" color translucent 0.45"
+See isosurface help of Jmol documentation.
+
 
 OPTIONS:
    -h      Show this message
    -i      Input filename.
+   -c      Jmolscript command
+   -d      Debug
 EOF
 }
 
-while getopts "hi:" OPTION
+while getopts "hdi:c:" OPTION
 do
      case $OPTION in
          h)
              usage
              exit 1
              ;;
+         d)
+             DEBUG="TRUE"
+             ;;
          i)
              FN=$OPTARG
+             ;;
+         c)
+             CMD=$OPTARG
              ;;
          ?)
              usage
@@ -72,7 +119,7 @@ fi
 
 
 #temporary jmolscript file for converting isourface in JVXL format
-sptfn=jmolscript_$$".spt"
+sptfn=jmolscript_$(((RANDOM%10000)+1000))".spt"
 
 # isosurface jvxl filename
 o=${FN%.*}".jvxl"
@@ -80,17 +127,24 @@ o=${FN%.*}".jvxl"
 xyz=${FN%.*}".xyz"
 # write script and do isosurface conversion
 echo "load \"$FN\"" > $sptfn
+if [[ -n $CMD ]]; then
+    echo $CMD >> $sptfn
+fi
 echo "write isosurface \"$o\"" >> $sptfn
 
 echo "Converting file $FN"
-/opt/jmol/jmol.sh -ions $sptfn
-rm $sptfn
+sync
+$Jmolexe -ions $sptfn
+
+if [[ $DEBUG != "TRUE" ]]; then
+    rm $sptfn
+fi
+
+sptfn=jmolscript_$(((RANDOM%10000)+1000))".spt"
+
 echo "load \"$FN\"" > $sptfn
-
-
 # Check if compressed
 type=`file ${FN} | grep zip`
-echo "type is:$type ."
 if [[ -z "$type" ]]; then
     echo "Proceed noncompressed data"
     CAT=cat
@@ -111,5 +165,8 @@ echo "load \"$xyz\"" > $sptfn
 echo "background [120,180,180]" >> $sptfn
 echo "isosurface s1 sign cyan yellow \"$o\" color translucent 0.45" >> $sptfn
 echo "write ZIP  \"${FN%.*}.zip\"" >> $sptfn
-/opt/jmol/jmol.sh -ions $sptfn
-rm  $sptfn $xyz $o
+sync
+$Jmolexe -ions $sptfn
+if [[ $DEBUG != "TRUE" ]]; then
+    rm  $sptfn $xyz $o
+fi
